Fine working code for sprint 1 (Admin, Superadmin and member roles. Create/Delete channel, Assign/Deassign, Promote/Demote functionality) Feb 15.

+++ removechannel.js
@@ -1,11 +1,17 @@
 import db from './Database-conf.js';
 
 async function removechannel(channelId) {
+
+    if (!channelId || typeof channelId !== 'string') {
+        console.error("Invalid channelId:", channelId);
+        return;
+    }
+
     try{
+
     const channel = await db.collection('channels').doc(channelId).delete();
 } catch (error) {
         console.error("Error deleting channel", error);
 }
 }
 export default removechannel;
-//console.log(await removechannel("Ti0ywCQ7ZbINW2uvw4od"));
